Work on inventory controller and script to add items
Added some work to the inventory controller object, as well as a script used to either add the item as a new row to the ds_grid, or add its amount to an existing item's amount.
This commit is done as a snapshot, as I am looking to remove variables as a value stored in the ds_grid and simply use the id of the items for all checks.