Fade Out script changed to Room Transition
The Fade Effect script has been changed and tailored to now handle going between rooms; all doors run this script, creating an object that handles all the fading out and in, as well as going between rooms and putting the player in the right coords.
Currently, moving the player to the right coords doesn't work.