Basic drawing item sprites to inventory screen
Very basic iteration of drawing items' inventory sprites to a mockup of the inventory screen, with a sprite drawn to a slot, iterating through the items in the inventory.
From here, need to actually make it look better, roughly, as well as streamline it a bit and draw a reticle over the item being selected.