Changes before 'preload' room is made
Determined I need to make a 'preload' room going forward, where important objects are instantiated and important values are set, before immediately after moving to the main menu.
The idea is to make the transition seem seamless, as though it never happened and instantly goes to the main menu.
This is likely where save data and setting would be imported before the main menu.